feat: content item 분리
